add rotas views e melhoria de estrutura

+++ app/Services/UserService.php
@@ -4,6 +4,7 @@ namespace App\Services;
 
 use App\Helper\ValidationHelper;
 use App\Repository\UserRepository;
+use Exception;
 use Illuminate\Support\Facades\Validator;
 
 class UserService 
@@ -17,6 +18,11 @@ class UserService
     }
 
     public function findById($id){
+
+        if(empty($id) || $id == ''){
+            throw new Exception('Id informado nao encontrado');
+        }
+
         return $this->repository->findById($id);
     }
 
@@ -32,14 +38,54 @@ class UserService
     }
 
     public function create($request){
-        return $this->repository->create($request);
+
+        $errors = ValidationHelper::validateRequest($request->all(), [
+            "name"     => "required",
+            "email"    => "email",
+            "password" => "required|string"
+        ]);
+
+        if(!empty($errors)){
+            return new Exception($errors);
+        }
+
+        $request['password'] = bcrypt($request['password']);
+
+        return $this->repository->create([
+            "name"     => $request["name"],
+            "email"    => $request["email"],
+            "password" => $request["password"],
+        ]);
     }
 
-    public function update($user,$request){
+    public function update($id,$request){
+
+        $errors = ValidationHelper::validateRequest($request->all(), [
+            "name"     => "required",
+            "email"    => "email",
+        ]);
+
+        if(!empty($errors)){
+            return new Exception($errors);
+        }
+
+        $user = $this->findById($id);
+
+        if(!$user){
+            return new Exception("Usuario informado nao foi localizado");
+        }
+
         return $this->repository->update($user,$request);
     }
 
-    public function delete($user){
+    public function delete($id){
+
+        $user = $this->findById($id);
+
+        if(!$user){
+            throw new Exception("Usuario informado nao foi localizado");
+        }
+
         return $this->repository->delete($user);
     }
 }
